% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tools.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Popup Label for Earthquake Map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{The \code{earthquakes} data frame.}
}
\value{
A vector with the HTML-formatted labels. Include this
vector with the data frame that is sent to \code{eq_map}.
}
\description{
Generates HTML labels for earthquake map
}
\details{
\code{eq_create_label} creates a more descriptive and HTML-formatted popup
label to be used in \code{\link{eq_map}}.

This function creates a vector of HTML-formatted labels using supplied
\code{earthquakes} data.  The function creates lines of the format
\strong{Label:} \code{ value} from the following variables in the
 \code{earthquakes} data set:
\itemize{
  \item Date_ymd
  \item Location_Name (as cleaned in the \code{eq_location_clean}
  function).
  \item EQ_Primary (earthquake magnitude)
  \item Total_Deaths
}
Any of the above variables with missing/NA values are omitted from the label.
}
\examples{

\dontrun{
earthquakes <- earthquakes \%>\%
 dplyr::filter(Country == 'China') \%>\%
 dplyr::filter(lubridate::year(Date_ymd) >= 2000) \%>\%
 dplyr::mutate(Popup_Text = eq_create_label(.))
 eq_map(earthquakes, annot_col = "Popup_Text")
}
}
