% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Data Preparation for Earthquakes Database}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{Earthquakes data frame from NOAA database.}
}
\description{
\code{eq_location_clean} function strips out country name and
converts column names to title case for readability.
}
\examples{
\dontrun{
earthquakes <- eq_location_clean(noaa_data)
}
}
