% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tools.R
\name{eq_map}
\alias{eq_map}
\title{Interactive Earthquake Map and Related Data}
\usage{
eq_map(data, annot_col = "Date_ymd")
}
\arguments{
\item{data}{The \code{earthquakes} data frame based on information taken from
the NOAA earthquake databases. Because there is a lot of data, the user should
  filter to a single country or an adjacent group of countries.}

\item{annot_col}{A column to use for the popup annotation. The user can select a
single column; otherwise you can create a more informative label first by using
the \code{eq_create_label} function.}
}
\value{
Returns an interactive \code{leaflet} map.
}
\description{
Generates interactive earthquake map and related data.
}
\details{
\code{eq_map} creates an interactive \code{leaflet} map
showing the location of earthquakes in the given \code{earthquakes} data set.

This function shows an interactive map of the location of the earthquakes in
the given \code{earthquakes} data.  The size of the circles are
proportional to the magnitude of the earthquakes (in the \code{EQ_PRIMARY})
variable. The map is interactive, and when you click on a link, the popup
shows the annotation as specified by the \code{annot_col} variable.
}
\examples{

## Using "Date_ymd" column as annotation
\dontrun{
earthquakes <- earthquakes \%>\%
   dplyr::filter(Country == 'China') \%>\%
   dplyr::filter(lubridate::year(Date_ymd) >= 2000)
eq_map(earthquakes, annot_col = 'Date_ymd')

earthquakes <- earthquakes \%>\%
 dplyr::filter(Country == 'China') \%>\%
 dplyr::filter(lubridate::year(Date_ymd) >= 2000) \%>\%
 dplyr::mutate(Popup_Text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "Popup_Text")
}
}
\seealso{
\code{eq_create_label} to create a more informative popup
  annotation in the \code{earthquakes} data frame.
}
