% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Generate annotations for points on map of earthquake data.}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes(). If specified and
inherit.aes = TRUE (the default), it is combined with the default mapping at
the top level of the plot.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transormation to use on the data for this layer,
as a string.}

\item{position}{Position adjustment, either as a string or the result of a
call to a position adjustment function.}

\item{na.rm}{If FALSE, missage values are removed with a warning.}

\item{show.legend}{Should this layer be included, any aesthetics are mapped.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining them.}

\item{...}{Other parameters to be included.}
}
\description{
\code{geom_timeline_label()} adds annotations to the earthquake data.
The geom adds a vertical line to each data point with text annotation providing
the location of the earthquake. This is a function that builds a layer, based
 on geom specification. It takes the geom definition and the coordinate data together.
}
