% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
ggproto object to display the earthquake time line
}
\keyword{datasets}
