% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimelineLabel
}
\description{
ggproto object to display the labels for earthquake time line
}
\keyword{datasets}
