% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaa_data}
\alias{noaa_data}
\title{Worldwide earthquake data from NOAA datebase}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1219 rows and 39 columns.
}
\source{
\url{https://data.noaa.gov/metaview/page?xml=NOAA/NESDIS/NGDC/MGG/Hazards/iso/xml/G012153.xml&view=getDataView#}
}
\usage{
data(noaa_data)
}
\description{
{National Geophysical Data Center / World Data Service (NGDC/WDS):
NCEI/WDS Global Significant Earthquake Database.
NOAA National Centers for Environmental Information.
doi:10.7289/V5TD9V7K}
}
\keyword{datasets}
