% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{\%>\%}
\alias{\%>\%}
\alias{eq_clean_data}
\title{Clean the row signif_dat}
\usage{
eq_clean_data(dat = read_signif())
}
\value{
a tibble containing 48 vaiables
}
\description{
A date column created by uniting the year, month, day and converting it to the Date class
    LATITUDE and LONGITUDE columns converted to numeric class
}
\examples{
clean_dat <- eq_clean_data()

}
