% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{titlecase}
\alias{titlecase}
\title{eq_location_clean() that cleans the LOCATION_NAME column by stripping out the country
name (including the colon) and converts names to title case (as opposed to all caps).}
\usage{
titlecase(x)
}
\value{
a data frame that cleans the LOCATION_NAME column by stripping out the country name (including the colon) and converts names to title case (as opposed to all caps).
}
\description{
eq_location_clean() that cleans the LOCATION_NAME column by stripping out the country
name (including the colon) and converts names to title case (as opposed to all caps).
}
\examples{
eq_location_clean()

}
