% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data() that takes raw NOAA data frame and returns a clean data frame.
The clean data frame should have the following:
A date column created by uniting the year, month, day and converting it to the Date class.
LATITUDE and LONGITUDE columns converted to numeric class}
\usage{
eq_clean_data()
}
\value{
a data frame created by uniting the year, month, day and converting it to the Date class. LATITUDE and LONGITUDE columns converted to numeric class
}
\description{
eq_clean_data() that takes raw NOAA data frame and returns a clean data frame.
The clean data frame should have the following:
A date column created by uniting the year, month, day and converting it to the Date class.
LATITUDE and LONGITUDE columns converted to numeric class
}
\examples{
eq_clean_data()

}
